pyproject.toml
[project]
name = "paddle-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paddlepaddle==3.2.0",
    "paddleocr==3.3.0",
    "fastapi>=0.135.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
